test(order): increase coverage order service
